requirements.txt
Django==3.2.5
psycopg2-binary==2.9.1
djangorestframework==3.12.4
djangorestframework-simplejwt
django-allauth==0.57.0
dj-rest-auth==5.0.1
urllib3<2.0.0
requests<3.0.0
djangorestframework-simplejwt==4.8.0
channels-redis
channels
daphne==4.1.2
